requirements.txt
apache-airflow>=2.0.0
google-cloud-storage>=2.0.0
kagglehub>=0.1.0
pendulum>=2.1.0